Move UserCsv-related files to csv, deprecated subdirs
