+++ src/deprecated/users-link.ts
@@ -1,4 +1,4 @@
-import { z } from 'zod';
+import * as z from 'zod';
 import { validateCsvData } from './csv';
 import { NormalizedUserTypeSchema } from './users';
